requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
chromadb==0.4.15

# AI/ML
openai==1.3.0
langchain==0.0.334
langchain-experimental==0.0.47
pandas==2.1.3
numpy==1.24.3

# Data Science
prophet==1.1.4
scikit-learn==1.3.2
plotly==5.17.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
sentence-transformers==2.2.2
requests==2.31.0